requirement.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
hiredis==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
Markdown==3.3.4
msgpack==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.2.0
